requirements.txt
python-decouple
opencv-python
python-telegram-bot[job-queue]
portalocker
aiohttp
asyncio